requirements.txt
flask
pillow
requests
beautifulsoup4
lxml
